refactor: split FSM cursor integration tests
